extract node 2d to core_pipeline
