+++ crates/bevy_gizmos/src/pipeline_2d.rs
@@ -1,23 +1,18 @@
 use bevy_asset::Handle;
-use bevy_core_pipeline::prelude::Camera2d;
+use bevy_core_pipeline::gizmo_2d::GizmoLine2d;
 use bevy_ecs::{
     prelude::Entity,
     query::With,
-    system::{Commands, Query, Res, ResMut, Resource},
+    system::{Query, Res, ResMut, Resource},
     world::{FromWorld, World},
 };
 use bevy_render::{
     mesh::{Mesh, MeshVertexBufferLayout},
-    prelude::Camera,
     render_asset::RenderAssets,
-    render_phase::{
-        CachedRenderPipelinePhaseItem, DrawFunctionId, DrawFunctions, PhaseItem, RenderPhase,
-        SetItemPipeline,
-    },
+    render_phase::{DrawFunctions, RenderPhase, SetItemPipeline},
     render_resource::*,
     texture::BevyDefault,
     view::Msaa,
-    Extract,
 };
 use bevy_sprite::*;
 use bevy_utils::FloatOrd;
@@ -96,57 +91,6 @@ pub(crate) type DrawGizmoLines = (
     SetMesh2dBindGroup<1>,
     DrawMesh2d,
 );
-
-pub struct GizmoLine2d {
-    pub sort_key: FloatOrd,
-    pub pipeline: CachedRenderPipelineId,
-    pub entity: Entity,
-    pub draw_function: DrawFunctionId,
-}
-
-impl PhaseItem for GizmoLine2d {
-    type SortKey = FloatOrd;
-
-    #[inline]
-    fn entity(&self) -> Entity {
-        self.entity
-    }
-
-    #[inline]
-    fn sort_key(&self) -> Self::SortKey {
-        self.sort_key
-    }
-
-    #[inline]
-    fn draw_function(&self) -> DrawFunctionId {
-        self.draw_function
-    }
-
-    #[inline]
-    fn sort(items: &mut [Self]) {
-        items.sort_by_key(|item| item.sort_key());
-    }
-}
-
-impl CachedRenderPipelinePhaseItem for GizmoLine2d {
-    #[inline]
-    fn cached_pipeline(&self) -> CachedRenderPipelineId {
-        self.pipeline
-    }
-}
-
-pub fn extract_gizmo_line_2d_camera_phase(
-    mut commands: Commands,
-    cameras_2d: Extract<Query<(Entity, &Camera), With<Camera2d>>>,
-) {
-    for (entity, camera) in &cameras_2d {
-        if camera.is_active {
-            commands
-                .get_or_spawn(entity)
-                .insert(RenderPhase::<GizmoLine2d>::default());
-        }
-    }
-}
 
 #[allow(clippy::too_many_arguments)]
 pub(crate) fn queue_gizmos_2d(
